backend/requirements.txt
flask
flask-cors
gunicorn
ultralytics
opencv-python-headless
numpy
torch
torchvision
google-generativeai